+++ header.php
@@ -43,7 +43,7 @@
 			</div>
 		</div>
 
-		<nav class="site-navigation top-bar" role="navigation">
+		<nav class="site-navigation top-bar" role="navigation" id="<?php foundationpress_mobile_menu_id(); ?>">
 			<div class="top-bar-left">
 				<div class="site-desktop-title top-bar-title">
 					<a href="<?php echo esc_url( home_url( '/' ) ); ?>" rel="home"><img class="desktop-site-logo" src="<?php echo bp_site_logo_src(); ?>" alt="Button Poetry Home"></a></a>