+++ template-parts/shop-page-products.php
@@ -1,0 +1,70 @@
+<?php
+/**
+ * Product Archive template part.
+ *
+ * This file mimics the default WooCommerce product archive template so it can be included on the shop page manually.
+ * 
+ * @package ButtonTheme
+ * @since ButtonTheme 1.3.0
+ */
+
+/**
+ * Hook: woocommerce_before_main_content.
+ *
+ * @hooked woocommerce_output_content_wrapper - 10 (outputs opening divs for the content)
+ * @hooked woocommerce_breadcrumb - 20
+ * @hooked WC_Structured_Data::generate_website_data() - 30
+ */
+?>
+
+<section class="main-container all-products-category-row">
+	<h2 class="text-center">All Button Products<?php //woocommerce_page_title(); ?></h2>
+	<div class="main-grid">
+		<main class="main-content">
+			<?php do_action( 'woocommerce_before_main_content' ); 
+			if ( woocommerce_product_loop() ) {
+				/**
+				 * Hook: woocommerce_before_shop_loop.
+				 *
+				 * @hooked woocommerce_output_all_notices - 10
+				 * @hooked woocommerce_result_count - 20
+				 * @hooked woocommerce_catalog_ordering - 30
+				 */
+				do_action( 'woocommerce_before_shop_loop' );
+				woocommerce_product_loop_start();
+				if ( wc_get_loop_prop( 'total' ) ) {
+					while ( have_posts() ) {
+						the_post();
+						/**
+						 * Hook: woocommerce_shop_loop.
+						 */
+						do_action( 'woocommerce_shop_loop' );
+						wc_get_template_part( 'content', 'product' );
+					}
+				}
+				woocommerce_product_loop_end();
+				/**
+				 * Hook: woocommerce_after_shop_loop.
+				 *
+				 * @hooked woocommerce_pagination - 10
+				 */
+				do_action( 'woocommerce_after_shop_loop' );
+			} else {
+				/**
+				 * Hook: woocommerce_no_products_found.
+				 *
+				 * @hooked wc_no_products_found - 10
+				 */
+				do_action( 'woocommerce_no_products_found' );
+			}
+			/**
+			 * Hook: woocommerce_after_main_content.
+			 *
+			 * @hooked woocommerce_output_content_wrapper_end - 10 (outputs closing divs for the content)
+			 */
+			do_action( 'woocommerce_after_main_content' );
+			?>
+		</main>
+		<?php get_sidebar(); ?>
+	</div>
+</section>