+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Button Theme Change Log
 For FoundationPress changes, refer to [FoundationPress on Github](https://github.com/olefredrik/FoundationPress/).
 
+## Next Release
+* Update: Re-synced with FoundationPress upstream repository
+
 ## 1.2.4
 ### 2019-05-22
 * Fixed: Updated WooCommerce email templates so they don't throw errors in WooCommerce 3.6.3.